# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API REST multi-session para WhatsApp usando whatsmeow. Baseada na WuzAPI com campos JSON em PascalCase.",
        "title": "FioZap API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "FioZap Support",
            "url": "https://github.com/fiozap/fiozap"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/sessions/{name}/contacts/check": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Verifica se os numeros informados estao no WhatsApp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Verificar numeros",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lista de numeros",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CheckPhoneRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ContactsCheckResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/contacts/{phone}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna informacoes de um contato (status, foto, devices)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Obter info do contato",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Numero do telefone",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.UserInfoResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/contacts/{phone}/avatar": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna URL da foto de perfil do contato",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Obter foto de perfil",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Numero do telefone",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.AvatarResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
        "/sessions/{name}/contacts/{phone}/business": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retorna perfil comercial do contato (nao implementado)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "contacts"
                ],
                "summary": "Obter perfil comercial",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome da sessao",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Numero do telefone",
                        "name": "phone",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.BusinessProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "501": {
                        "description": "Not Implemented",
                        "schema": {
                            "$ref": "#/definitions/dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.AvatarResponse": {
            "type": "object",
            "properties": {
                "ID": {
                    "type": "string",
                    "example": "122345678901234567890"
                },
                "URL": {
                    "type": "string",
                    "example": "https://pps.whatsapp.net/..."
                }
            }
        },
...
        "dto.BusinessProfileResponse": {
            "type": "object",
            "properties": {
                "Address": {
                    "type": "string"
                },
                "Category": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "JID": {
                    "type": "string"
                },
                "Website": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "dto.CheckPhoneRequest": {
            "type": "object",
            "properties": {
                "Phone": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "5511999999999",
                        "5521888888888"
                    ]
                }
            }
        },
        "dto.CheckPhoneResponse": {
            "type": "object",
            "properties": {
                "IsInWhatsapp": {
                    "type": "boolean",
                    "example": true
                },
                "JID": {
                    "type": "string",
                    "example": "5511999999999@s.whatsapp.net"
                },
                "Query": {
                    "type": "string",
                    "example": "5511999999999"
                }
            }
        },
        "dto.ContactsCheckResponse": {
            "type": "object",
            "properties": {
                "Contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CheckPhoneResponse"
                    }
                }
            }
        },
...
        "dto.Response": {
            "description": "Resposta padrao da API FioZap",
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "dto.UserInfoResponse": {
            "type": "object",
            "properties": {
                "Devices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "JID": {
                    "type": "string",
                    "example": "5511999999999@s.whatsapp.net"
                },
                "PictureID": {
                    "type": "string"
                },
                "Status": {
                    "type": "string",
                    "example": "Hey there! I am using WhatsApp."
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Token global (GLOBAL_API_TOKEN) ou token da sessao",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}